% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{useGA}
\alias{useGA}
\title{Use Google Analytics}
\usage{
useGA(id)
}
\arguments{
\item{id}{The Google Analytics tracking ID.}
}
\description{
Add Google Analytics to your Shiny app
}
\examples{
if(FALSE){
library(shiny)

ui <- navbarPage(
  "Google Analytics",
  header = list(
    useGA("G-59ZYN9HXFQ")
  ),
  tabPanel(
    "Inputs",
    textInput("name", "Name"),
    selectInput("sex", "Sex", c("Male", "Female")),
    radioButtons("color", "Color", c("Red", "Green", "Blue"))
  ),
  tabPanel(
    "Outputs",
    selectizeInput("cities", "Cities", choices = c("New York", "Los Angeles", "Chicago")),
    checkboxGroupInput("fruits", "Fruits", choices = c("Apple", "Banana", "Orange")),
    actionButton("submit", "Submit"),
    plotOutput("plot")
  )
)

server <- function(input, output, session) {
  output$plot <- renderPlot({
    input$submit
    plot(runif(10))
  })
}

shinyApp(ui, server)
}
}
